Handle signals manually instead of using ctrlc
This is part of size optimization. ctrlc includes extra logic to allow
invoking non-signal-safe functions in response to an interrupt signal.
Internally this spawns threads and adds some global state to track a
semaphore that the unsafe signal handler unlocks so that safe code may
continue.

We don't need this because killing the child process is safe to do
within the signal handler. Getting rid of ctrlc takes us from 404k to
364k without any loss of functionality.